+++ packages/mcp/src/auth/middleware/index.ts
@@ -1,0 +1,1 @@
+export * from "./clientAuth";
